feat(facebook): import detects token death + dedups via external_id
- Replace string-match rate-limit catch with typed FacebookTransientException / FacebookTokenInvalidException handlers; dead token sets connection status to NeedsReauth, records last_error, fires FacebookConnectionNeedsReauth event, and marks source as Error
- Replace whereJsonContains raw_data duplicate check with external_id column lookup for reliable idempotency; set external_id on create and in updateExisting branch
- Remove redundant outer (int) cast from Lead::nextSortForPhase (closure already returns int)
- Add ImportFacebookLeadsSelfHealTest covering token-death → NeedsReauth/event/source-Error and duplicate-lead prevention

+++ src/Models/Lead.php
@@ -62,7 +62,7 @@ class Lead extends Model
      */
     public static function nextSortForPhase(int|string $phaseKey): int
     {
-        return (int) DB::transaction(function () use ($phaseKey): int {
+        return DB::transaction(function () use ($phaseKey): int {
             $max = static::query()
                 ->where(static::fkColumn('lead_phase'), $phaseKey)
                 ->lockForUpdate()
